requirements.txt
Flask
connexion
Flask-Injector
Flask_Cors
fastavro
elasticsearch
pyjwt